+++ app/src/views/team/[id]/ExpenseAccountView.vue
@@ -61,20 +61,20 @@
         <AddressToolTip :address="team?.expenseAccountEip712Address ?? ''" class="text-xs" />
       </div>
     </div>
-    <div
+    <!--<div
       v-if="team?.expenseAccountEip712Address"
       class="card shadow-xl bg-white flex text-primary-content p-5 overflow-visible mb-10"
     >
       <span class="text-2xl font-bold">My Approved Expense</span>
-      <!-- TODO display this only if the use have an approved expense -->
-      <!-- Expense A/c Info Section -->
+      -- TODO display this only if the use have an approved expense --
+      -- Expense A/c Info Section --
       <section class="stat flex flex-col justify-start">
-        <!-- New Header -->
+        -- New Header --
 
         <div class="flex flex-col gap-8">
           <div class="overflow-x-auto" data-test="approval-table">
             <table class="table">
-              <!-- head -->
+              -- head --
               <thead class="text-sm font-bold">
                 <tr>
                   <th>Expiry Date</th>
@@ -147,7 +147,15 @@
           />
         </ModalComponent>
       </section>
-    </div>
+    </div>-->
+    <MyApprovedExpenseSection 
+      v-if="team" 
+      :team="team" 
+      :is-disapproved-address="isDisapprovedAddress"
+      :token-symbol="tokenSymbol"
+      :expense-balance-formatted="expenseBalanceFormatted"
+      :usdc-balance="usdcBalance"
+    />
 
     <div
       class="card shadow-xl bg-white p-5 overflow-x-auto flex flex-col gap-4"
@@ -179,6 +187,18 @@
       />
       <!--</div>
       </div>-->
+        <ModalComponent v-model="approveUsersModal">
+          <ApproveUsersForm
+            v-if="approveUsersModal"
+            :form-data="teamMembers"
+            :users="foundUsers"
+            :loading-approve="loadingApprove"
+            :is-bod-action="isBodAction()"
+            @approve-user="approveUser"
+            @close-modal="approveUsersModal = false"
+            @search-users="(input) => searchUsers(input)"
+          />
+        </ModalComponent>
     </div>
 
     <div
@@ -227,6 +247,7 @@ import ERC20ABI from '@/artifacts/abi/erc20.json'
 import { readContract } from '@wagmi/core'
 import { config } from '@/wagmi.config'
 import { useRoute } from 'vue-router'
+import MyApprovedExpenseSection from '@/components/sections/ExpenseAccountView/MyApprovedExpenseSection.vue'
 //#endregion
 
 //#region Refs
@@ -248,32 +269,32 @@ const deactivateIndex = ref<number | null>(null)
 const expenseAccountEip712Address = computed(
   () => team.value?.expenseAccountEip712Address as Address
 )
-const maxLimit = (budgetType: number) =>
-  computed(() => {
-    const budgetData =
-      _expenseAccountData.value?.data &&
-      Array.isArray(JSON.parse(_expenseAccountData.value?.data).budgetData)
-        ? JSON.parse(_expenseAccountData.value?.data).budgetData.find(
-            (item: BudgetData) => item.budgetType === budgetType
-          )
-        : undefined
-    if (_expenseAccountData.value?.data && budgetData && budgetData.budgetType === budgetType)
-      return budgetData.value
-    else return '--'
-  })
-const maxLimitTxsPerPeriod = maxLimit(0)
-const maxLimitAmountPerPeriod = maxLimit(1)
-const maxLimitAmountPerTx = maxLimit(2)
+// const maxLimit = (budgetType: number) =>
+//   computed(() => {
+//     const budgetData =
+//       _expenseAccountData.value?.data &&
+//       Array.isArray(JSON.parse(_expenseAccountData.value?.data).budgetData)
+//         ? JSON.parse(_expenseAccountData.value?.data).budgetData.find(
+//             (item: BudgetData) => item.budgetType === budgetType
+//           )
+//         : undefined
+//     if (_expenseAccountData.value?.data && budgetData && budgetData.budgetType === budgetType)
+//       return budgetData.value
+//     else return '--'
+//   })
+// const maxLimitTxsPerPeriod = maxLimit(0)
+// const maxLimitAmountPerPeriod = maxLimit(1)
+// const maxLimitAmountPerTx = maxLimit(2)
 
-const expiry = computed(() => {
-  if (_expenseAccountData.value?.data) {
-    const unixEpoch = JSON.parse(_expenseAccountData.value.data).expiry
-    const date = new Date(Number(unixEpoch) * 1000)
-    return date.toLocaleString('en-US')
-  } else {
-    return '--/--/--, --:--:--'
-  }
-})
+// const expiry = computed(() => {
+//   if (_expenseAccountData.value?.data) {
+//     const unixEpoch = JSON.parse(_expenseAccountData.value.data).expiry
+//     const date = new Date(Number(unixEpoch) * 1000)
+//     return date.toLocaleString('en-US')
+//   } else {
+//     return '--/--/--, --:--:--'
+//   }
+// })
 const tokenSymbol = (tokenAddress: string) =>
   computed(() => {
     const symbols = {
@@ -284,45 +305,45 @@ const tokenSymbol = (tokenAddress: string) =>
 
     return symbols[tokenAddress] || ''
   })
-const dynamicDisplayData = (budgetType: number) =>
-  computed(() => {
-    const data =
-      budgetType === 0
-        ? { heading: 'Total Transactions', symbol: 'TXs' }
-        : { heading: 'Total Withdrawn', symbol: NETWORK.currencySymbol }
-    if (_expenseAccountData.value?.data && amountWithdrawn.value) {
-      if (budgetType === 0) {
-        return {
-          ...data,
-          // @ts-expect-error: amountWithdrawn.value is a array
-          value: Number(amountWithdrawn.value[0])
-        }
-      } else {
-        const tokenAddress = JSON.parse(_expenseAccountData.value.data).tokenAddress
-        return {
-          ...data,
-          value:
-            tokenAddress === zeroAddress
-              ? // @ts-expect-error: amountWithdrawn.value is a array
-                formatEther(amountWithdrawn.value[1])
-              : // @ts-expect-error: amountWithdrawn.value is a array
-                Number(amountWithdrawn.value[1]) / 1e6
-        }
-      }
-    } else {
-      return {
-        ...data,
-        value: `--`
-      }
-    }
-  })
+// const dynamicDisplayData = (budgetType: number) =>
+//   computed(() => {
+//     const data =
+//       budgetType === 0
+//         ? { heading: 'Total Transactions', symbol: 'TXs' }
+//         : { heading: 'Total Withdrawn', symbol: NETWORK.currencySymbol }
+//     if (_expenseAccountData.value?.data && amountWithdrawn.value) {
+//       if (budgetType === 0) {
+//         return {
+//           ...data,
+//           // @ts-expect-error: amountWithdrawn.value is a array
+//           value: Number(amountWithdrawn.value[0])
+//         }
+//       } else {
+//         const tokenAddress = JSON.parse(_expenseAccountData.value.data).tokenAddress
+//         return {
+//           ...data,
+//           value:
+//             tokenAddress === zeroAddress
+//               ? // @ts-expect-error: amountWithdrawn.value is a array
+//                 formatEther(amountWithdrawn.value[1])
+//               : // @ts-expect-error: amountWithdrawn.value is a array
+//                 Number(amountWithdrawn.value[1]) / 1e6
+//         }
+//       }
+//     } else {
+//       return {
+//         ...data,
+//         value: `--`
+//       }
+//     }
+//   })
 const expenseBalanceFormatted = computed(() => {
   if (typeof expenseAccountBalance.value?.value === 'bigint')
     return formatEther(expenseAccountBalance.value.value)
   else return '--'
 })
-const dynamicDisplayDataTx = dynamicDisplayData(0)
-const dynamicDisplayDataAmount = dynamicDisplayData(1)
+// const dynamicDisplayDataTx = dynamicDisplayData(0)
+// const dynamicDisplayDataAmount = dynamicDisplayData(1)
 // Reactive storage for records to be displayed in table
 const manyExpenseAccountDataAll = reactive<ManyExpenseWithBalances[]>([])
 
@@ -346,24 +367,23 @@ const {
   .get()
   .json<Team>()
 
-const {
-  error: fetchExpenseAccountDataError,
-  // isFetching: isFetchingExpenseAccountData,
-  execute: fetchExpenseAccountData,
-  data: _expenseAccountData
-} = useCustomFetch(`teams/${String(route.params.id)}/expense-data`, {
-  immediate: false,
-  beforeFetch: async ({ options, url, cancel }) => {
-    options.headers = {
-      memberaddress: currentUserAddress,
-      'Content-Type': 'application/json',
-      ...options.headers
-    }
-    return { options, url, cancel }
-  }
-})
-  .get()
-  .json()
+// const {
+//   error: fetchExpenseAccountDataError,
+//   execute: fetchExpenseAccountData,
+//   data: _expenseAccountData
+// } = useCustomFetch(`teams/${String(route.params.id)}/expense-data`, {
+//   immediate: false,
+//   beforeFetch: async ({ options, url, cancel }) => {
+//     options.headers = {
+//       memberaddress: currentUserAddress,
+//       'Content-Type': 'application/json',
+//       ...options.headers
+//     }
+//     return { options, url, cancel }
+//   }
+// })
+//   .get()
+//   .json()
 
 const {
   error: fetchManyExpenseAccountDataError,
@@ -439,30 +459,30 @@ const {
   args: [expenseAccountEip712Address as unknown as Address]
 })
 
-const {
-  data: amountWithdrawn,
-  refetch: executeGetAmountWithdrawn,
-  error: errorGetAmountWithdrawn
-  //isLoading: isLoadingGetAmountWithdrawn
-} = useReadContract({
-  functionName: 'balances',
-  address: expenseAccountEip712Address as unknown as Address,
-  abi: expenseAccountABI,
-  args: [signatureHash]
-})
+// const {
+//   data: amountWithdrawn,
+//   refetch: executeGetAmountWithdrawn,
+//   error: errorGetAmountWithdrawn
+//   //isLoading: isLoadingGetAmountWithdrawn
+// } = useReadContract({
+//   functionName: 'balances',
+//   address: expenseAccountEip712Address as unknown as Address,
+//   abi: expenseAccountABI,
+//   args: [signatureHash]
+// })
 
 //expense account transfer
-const {
-  writeContract: executeExpenseAccountTransfer,
-  isPending: isLoadingTransfer,
-  error: errorTransfer,
-  data: transferHash
-} = useWriteContract()
+// const {
+//   writeContract: executeExpenseAccountTransfer,
+//   isPending: isLoadingTransfer,
+//   error: errorTransfer,
+//   data: transferHash
+// } = useWriteContract()
 
-const { isLoading: isConfirmingTransfer, isSuccess: isConfirmedTransfer } =
-  useWaitForTransactionReceipt({
-    hash: transferHash
-  })
+// const { isLoading: isConfirmingTransfer, isSuccess: isConfirmedTransfer } =
+//   useWaitForTransactionReceipt({
+//     hash: transferHash
+//   })
 
 //deactivate approval
 const {
@@ -491,17 +511,17 @@ const { isLoading: isConfirmingActivate, isSuccess: isConfirmedActivate } =
   })
 
 // Token approval
-const {
-  writeContract: approve,
-  error: approveError,
-  data: approveHash
-  // isPending: isPendingApprove
-} = useWriteContract()
+// const {
+//   writeContract: approve,
+//   error: approveError,
+//   data: approveHash
+//   // isPending: isPendingApprove
+// } = useWriteContract()
 
-const { isLoading: isConfirmingApprove, isSuccess: isConfirmedApprove } =
-  useWaitForTransactionReceipt({
-    hash: approveHash
-  })
+// const { isLoading: isConfirmingApprove, isSuccess: isConfirmedApprove } =
+//   useWaitForTransactionReceipt({
+//     hash: approveHash
+//   })
 //#endregion
 
 //#region Functions
@@ -543,11 +563,12 @@ const initializeBalances = async () => {
 
 const init = async () => {
   await getExpenseAccountOwner()
-  await fetchExpenseAccountData()
+  // await fetchExpenseAccountData()
   await fetchManyExpenseAccountData()
   await initializeBalances()
   await executeFetchTeam()
-  await getAmountWithdrawnBalance()
+  await executeGetExpenseAccountBalance()
+  // await getAmountWithdrawnBalance()
   await fetchUsdcBalance()
 }
 
@@ -579,101 +600,101 @@ const getExpenseAccountOwner = async () => {
   if (team.value?.expenseAccountEip712Address) await executeExpenseAccountGetOwner()
 }
 
-const getAmountWithdrawnBalance = async () => {
-  if (team.value?.expenseAccountEip712Address) {
-    if (!_expenseAccountData?.value?.data) return
-    signatureHash.value = keccak256(_expenseAccountData.value.signature)
-    await executeGetAmountWithdrawn()
-  }
-}
+// const getAmountWithdrawnBalance = async () => {
+//   if (team.value?.expenseAccountEip712Address) {
+//     if (!_expenseAccountData?.value?.data) return
+//     signatureHash.value = keccak256(_expenseAccountData.value.signature)
+//     await executeGetAmountWithdrawn()
+//   }
+// }
 
-const transferFromExpenseAccount = async (to: string, amount: string) => {
-  tokenAmount.value = amount
-  tokenRecipient.value = to
+// const transferFromExpenseAccount = async (to: string, amount: string) => {
+//   tokenAmount.value = amount
+//   tokenRecipient.value = to
 
-  if (team.value?.expenseAccountEip712Address && _expenseAccountData.value.data) {
-    const budgetLimit: BudgetLimit = JSON.parse(_expenseAccountData.value.data)
+//   if (team.value?.expenseAccountEip712Address && _expenseAccountData.value.data) {
+//     const budgetLimit: BudgetLimit = JSON.parse(_expenseAccountData.value.data)
 
-    if (budgetLimit.tokenAddress === zeroAddress) transferNativeToken(to, amount, budgetLimit)
-    else await transferErc20Token()
-  }
-}
+//     if (budgetLimit.tokenAddress === zeroAddress) transferNativeToken(to, amount, budgetLimit)
+//     else await transferErc20Token()
+//   }
+// }
 
-const transferNativeToken = (to: string, amount: string, budgetLimit: BudgetLimit) => {
-  executeExpenseAccountTransfer({
-    address: team.value?.expenseAccountEip712Address as Address,
-    args: [
-      to,
-      parseEther(amount),
-      {
-        ...budgetLimit,
-        budgetData: budgetLimit.budgetData.map((item) => ({
-          ...item,
-          value:
-            item.budgetType === 0
-              ? item.value
-              : budgetLimit.tokenAddress === zeroAddress
-                ? parseEther(`${item.value}`)
-                : BigInt(Number(item.value) * 1e6)
-        }))
-      },
-      _expenseAccountData.value.signature
-    ],
-    abi: expenseAccountABI,
-    functionName: 'transfer'
-  })
-}
+// const transferNativeToken = (to: string, amount: string, budgetLimit: BudgetLimit) => {
+//   executeExpenseAccountTransfer({
+//     address: team.value?.expenseAccountEip712Address as Address,
+//     args: [
+//       to,
+//       parseEther(amount),
+//       {
+//         ...budgetLimit,
+//         budgetData: budgetLimit.budgetData.map((item) => ({
+//           ...item,
+//           value:
+//             item.budgetType === 0
+//               ? item.value
+//               : budgetLimit.tokenAddress === zeroAddress
+//                 ? parseEther(`${item.value}`)
+//                 : BigInt(Number(item.value) * 1e6)
+//         }))
+//       },
+//       _expenseAccountData.value.signature
+//     ],
+//     abi: expenseAccountABI,
+//     functionName: 'transfer'
+//   })
+// }
 
-// Token transfer function
-const transferErc20Token = async () => {
-  if (
-    !team.value?.expenseAccountEip712Address ||
-    !tokenAmount.value ||
-    !tokenRecipient.value ||
-    !_expenseAccountData.value?.data
-  )
-    return
+// // Token transfer function
+// const transferErc20Token = async () => {
+//   if (
+//     !team.value?.expenseAccountEip712Address ||
+//     !tokenAmount.value ||
+//     !tokenRecipient.value ||
+//     !_expenseAccountData.value?.data
+//   )
+//     return
 
-  const tokenAddress = USDC_ADDRESS
-  const _amount = BigInt(Number(tokenAmount.value) * 1e6)
+//   const tokenAddress = USDC_ADDRESS
+//   const _amount = BigInt(Number(tokenAmount.value) * 1e6)
 
-  const budgetLimit: BudgetLimit = JSON.parse(_expenseAccountData.value.data)
+//   const budgetLimit: BudgetLimit = JSON.parse(_expenseAccountData.value.data)
 
-  const allowance = await readContract(config, {
-    address: tokenAddress as Address,
-    abi: ERC20ABI,
-    functionName: 'allowance',
-    args: [currentUserAddress as Address, team.value.expenseAccountEip712Address as Address]
-  })
+//   const allowance = await readContract(config, {
+//     address: tokenAddress as Address,
+//     abi: ERC20ABI,
+//     functionName: 'allowance',
+//     args: [currentUserAddress as Address, team.value.expenseAccountEip712Address as Address]
+//   })
 
-  const currentAllowance = allowance ? allowance.toString() : 0n
-  if (Number(currentAllowance) < Number(_amount)) {
-    approve({
-      address: tokenAddress as Address,
-      abi: ERC20ABI,
-      functionName: 'approve',
-      args: [team.value.expenseAccountEip712Address as Address, _amount]
-    })
-  } else {
-    executeExpenseAccountTransfer({
-      address: team.value.expenseAccountEip712Address as Address,
-      abi: expenseAccountABI,
-      functionName: 'transfer',
-      args: [
-        tokenRecipient.value as Address,
-        _amount,
-        {
-          ...budgetLimit,
-          budgetData: budgetLimit.budgetData.map((item) => ({
-            ...item,
-            value: item.budgetType === 0 ? item.value : BigInt(Number(item.value) * 1e6) //parseEther(`${item.value}`)
-          }))
-        },
-        _expenseAccountData.value.signature
-      ]
-    })
-  }
-}
+//   const currentAllowance = allowance ? allowance.toString() : 0n
+//   if (Number(currentAllowance) < Number(_amount)) {
+//     approve({
+//       address: tokenAddress as Address,
+//       abi: ERC20ABI,
+//       functionName: 'approve',
+//       args: [team.value.expenseAccountEip712Address as Address, _amount]
+//     })
+//   } else {
+//     executeExpenseAccountTransfer({
+//       address: team.value.expenseAccountEip712Address as Address,
+//       abi: expenseAccountABI,
+//       functionName: 'transfer',
+//       args: [
+//         tokenRecipient.value as Address,
+//         _amount,
+//         {
+//           ...budgetLimit,
+//           budgetData: budgetLimit.budgetData.map((item) => ({
+//             ...item,
+//             value: item.budgetType === 0 ? item.value : BigInt(Number(item.value) * 1e6) //parseEther(`${item.value}`)
+//           }))
+//         },
+//         _expenseAccountData.value.signature
+//       ]
+//     })
+//   }
+// }
 
 const approveUser = async (data: BudgetLimit) => {
   loadingApprove.value = true
@@ -749,15 +770,15 @@ watch(
   }
 )
 
-watch(isConfirmingTransfer, async (isConfirming, wasConfirming) => {
-  if (!isConfirming && wasConfirming && isConfirmedTransfer.value) {
-    addSuccessToast('Transfer Successful')
-    await executeGetExpenseAccountBalance()
-    await fetchUsdcBalance()
-    await getAmountWithdrawnBalance()
-    transferModal.value = false
-  }
-})
+// watch(isConfirmingTransfer, async (isConfirming, wasConfirming) => {
+//   if (!isConfirming && wasConfirming && isConfirmedTransfer.value) {
+//     addSuccessToast('Transfer Successful')
+//     await executeGetExpenseAccountBalance()
+//     await fetchUsdcBalance()
+//     await getAmountWithdrawnBalance()
+//     transferModal.value = false
+//   }
+// })
 
 watch(isConfirmingActivate, async (isConfirming, wasConfirming) => {
   if (!isConfirming && wasConfirming && isConfirmedActivate.value) {
@@ -796,19 +817,19 @@ watch(errorGetOwner, (newVal) => {
   if (newVal) addErrorToast(errorMessage(newVal, 'Error Getting Contract Owner'))
 })
 
-watch(errorGetAmountWithdrawn, (newVal) => {
-  if (newVal) {
-    log.error(parseError(newVal))
-    addErrorToast('Failed to fetch amount withdrawn')
-  }
-})
+// watch(errorGetAmountWithdrawn, (newVal) => {
+//   if (newVal) {
+//     log.error(parseError(newVal))
+//     addErrorToast('Failed to fetch amount withdrawn')
+//   }
+// })
 
-watch(errorTransfer, (newVal) => {
-  if (newVal) {
-    log.error(parseError(newVal))
-    addErrorToast('Failed to transfer')
-  }
-})
+// watch(errorTransfer, (newVal) => {
+//   if (newVal) {
+//     log.error(parseError(newVal))
+//     addErrorToast('Failed to transfer')
+//   }
+// })
 
 watch(errorDeactivateApproval, (newVal) => {
   if (newVal) {
@@ -839,9 +860,9 @@ watch(fetchManyExpenseAccountDataError, (newVal) => {
   }
 })
 
-watch(fetchExpenseAccountDataError, (newVal) => {
-  if (newVal) addErrorToast('Error fetching expense account data')
-})
+// watch(fetchExpenseAccountDataError, (newVal) => {
+//   if (newVal) addErrorToast('Error fetching expense account data')
+// })
 
 watch(isErrorExpenseAccountBalance, (newVal) => {
   if (newVal) {
@@ -850,19 +871,19 @@ watch(isErrorExpenseAccountBalance, (newVal) => {
   }
 })
 
-watch(isConfirmingApprove, (newIsConfirming, oldIsConfirming) => {
-  if (!newIsConfirming && oldIsConfirming && isConfirmedApprove.value) {
-    addSuccessToast('Approval granted successfully')
-    transferErc20Token()
-  }
-})
+// watch(isConfirmingApprove, (newIsConfirming, oldIsConfirming) => {
+//   if (!newIsConfirming && oldIsConfirming && isConfirmedApprove.value) {
+//     addSuccessToast('Approval granted successfully')
+//     transferErc20Token()
+//   }
+// })
 
-watch(approveError, () => {
-  if (approveError.value) {
-    log.error(parseError(approveError.value))
-    addErrorToast('Failed to approve token spending')
-  }
-})
+// watch(approveError, () => {
+//   if (approveError.value) {
+//     log.error(parseError(approveError.value))
+//     addErrorToast('Failed to approve token spending')
+//   }
+// })
 
 watch([usdcBalanceError], ([newUsdcError]) => {
   if (newUsdcError) {
